mutualize zstd checker
